# openapi.yaml
openapi: 3.0.3
info:
  title: Personal Investment Management API
  version: 0.2.0
  description: MVP API for personal investment management tool (snapshot-authoritative).
servers:
- url: '{{baseUrl}}'
  description: Variable base URL
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        institution_id:
          type: integer
          nullable: true
        product_code:
          type: string
          nullable: true
        product_type:
          type: string
          enum:
          - bank_wmp
          - money_market
          - term_deposit
          - fund
          - stock
          - other
        risk_level:
          type: string
          nullable: true
          description: External label, e.g. R1-R5
        term_days:
          type: integer
          nullable: true
        liquidity_rule:
          type: string
          enum:
          - open
          - closed
          - periodic_open
        settle_days:
          type: integer
        note:
          type: string
          nullable: true
        valuation_mode:
          type: string
          enum:
          - product_value
          - lot_value
          default: product_value
      required:
      - id
      - name
      - product_type
      - liquidity_rule
      - settle_days
      - valuation_mode
...
    ProductMetrics:
      type: object
      nullable: true
      properties:
        twr:
          type: number
          description: Time-Weighted Return (百分比，如 5.5 表示 5.5%)
        annualized:
          type: number
          description: Annualized return (百分比，如 8.2 表示 8.2%)
        volatility:
          type: number
          description: Annualized volatility (百分比，如 12.5 表示 12.5%)
        max_drawdown:
          type: number
          description: Maximum drawdown (百分比，如 -15.3 表示 -15.3%)
        drawdown_recovery_days:
          type: integer
          description: Days to recover from max drawdown
    ChartPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        market_value:
          type: number
        source:
          type: string
          enum:
          - manual
          - interpolated
          - extrapolated
          description: |
            估值点来源类型：
            - manual: 用户录入的真实估值点
            - interpolated: 在两个 manual 点之间线性插值
            - extrapolated: 在最后一个 manual 点之后外推（保持最后一个值）
      required:
      - date
      - market_value
      - source
...
paths:
...
  /api/products:
    get:
      summary: List products with optional metrics
      parameters:
      - name: include_metrics
        in: query
        required: false
        schema:
          type: boolean
          default: false
        description: 是否包含各窗口期的收益指标
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          allOf:
                          - $ref: '#/components/schemas/Product'
                          - type: object
                            properties:
                              total_principal:
                                type: number
                                description: 总持仓本金
                              total_shares:
                                type: number
                                description: 总持仓份额
                              latest_market_value:
                                type: number
                                description: 最新市值
                              metrics_by_window:
                                type: object
                                additionalProperties:
                                  $ref: '#/components/schemas/ProductMetrics'
                              metrics:
                                $ref: '#/components/schemas/ProductMetrics'
                  message:
                    type: string
                required:
                - code
                - data
                - message
    post:
      summary: Create product
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    $ref: '#/components/schemas/Product'
                  message:
                    type: string
                required:
                - code
                - data
                - message
              example:
                code: 200
                data:
                  id: 201
                  name: 日日金
                  institution_id: 1
                  product_code: null
                  product_type: money_market
                  risk_level: R1
                  term_days: 0
                  liquidity_rule: open
                  settle_days: 0
                  note: 随时申赎
                  valuation_mode: product_value
                message: ok
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                institution_id:
                  type: integer
                  nullable: true
                product_code:
                  type: string
                  nullable: true
                product_type:
                  type: string
                  enum:
                  - bank_wmp
                  - money_market
                  - term_deposit
                  - fund
                  - stock
                  - other
                risk_level:
                  type: string
                  nullable: true
                term_days:
                  type: integer
                  nullable: true
                liquidity_rule:
                  type: string
                  enum:
                  - open
                  - closed
                  - periodic_open
                settle_days:
                  type: integer
                note:
                  type: string
                  nullable: true
                valuation_mode:
                  type: string
                  enum:
                  - product_value
                  - lot_value
                  default: product_value
              required:
              - name
              - product_type
              - liquidity_rule
            example:
              name: 日日金
              institution_id: 1
              product_type: money_market
              risk_level: R1
              term_days: 0
              liquidity_rule: open
              settle_days: 0
              note: 随时申赎
  /api/products/{id}:
    patch:
      summary: Update product
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    $ref: '#/components/schemas/Product'
                  message:
                    type: string
                required:
                - code
                - data
                - message
              example:
                code: 200
                data:
                  id: 201
                  name: 日日金
                  institution_id: 1
                  product_code: null
                  product_type: money_market
                  risk_level: R1
                  term_days: 0
                  liquidity_rule: open
                  settle_days: 0
                  note: 随时申赎
                  valuation_mode: product_value
                message: ok
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                institution_id:
                  type: integer
                  nullable: true
                product_code:
                  type: string
                  nullable: true
                product_type:
                  type: string
                  enum:
                  - bank_wmp
                  - money_market
                  - term_deposit
                  - fund
                  - stock
                  - other
                risk_level:
                  type: string
                  nullable: true
                term_days:
                  type: integer
                  nullable: true
                liquidity_rule:
                  type: string
                  enum:
                  - open
                  - closed
                  - periodic_open
                settle_days:
                  type: integer
                note:
                  type: string
                  nullable: true
                valuation_mode:
                  type: string
                  enum:
                  - product_value
                  - lot_value
            example:
              note: 更新备注
    delete:
      summary: Delete product
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  message:
                    type: string
                required:
                - code
                - data
                - message
              example:
                code: 200
                data:
                  message: deleted
                message: ok
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                required:
                - code
                - message
              example:
                code: 404
                message: Product 201 not found
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
  /api/products/{id}/chart:
    get:
      summary: Get product valuation chart data
      description: 获取产品市值走势，返回 manual 和 interpolated 点
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: window
        in: query
        required: false
        schema:
          type: string
          enum:
          - 4w
          - 8w
          - 12w
          - 24w
          - 1y
          - ytd
          default: 8w
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      product_id:
                        type: integer
                      valuation_mode:
                        type: string
                      points:
                        type: array
                        items:
                          $ref: '#/components/schemas/ChartPoint'
                  message:
                    type: string
                required:
                - code
                - data
                - message
...